Add MenuBar to each pages
