+++ src/pages/Difficulty.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import MenuBar from '../components/MenuBar';
 
 function Difficulty() {
   return (
     <div>
+      <MenuBar />
       <h1>지역별 운전 난이도</h1>
     </div>
   );
