+++ src/pages/Home.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import MenuBar from '../components/MenuBar';
 
 function Home() {
   return (
     <div>
+      <MenuBar />
       <h1>홈 페이지</h1>
     </div>
   );
