+++ src/pages/Login.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import MenuBar from '../components/MenuBar';
 
 function Login() {
   return (
     <div>
+      <MenuBar />
       <h1>로그인</h1>
     </div>
   );
