+++ src/pages/Ranking.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import MenuBar from '../components/MenuBar';
 
 function Ranking() {
   return (
     <div>
+      <MenuBar />
       <h1>운전자 랭킹</h1>
     </div>
   );
